backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
websockets==12.0
google-generativeai==0.8.3
python-dotenv==1.0.1
aiohttp==3.9.3
pydantic==2.6.3
openai-whisper
torch
numpy>=1.25.2
pyannote.audio
speechbrain

